Organize settings window views by sections
