+++ Loop/Settings Window/Loop/AdvancedConfiguration.swift
@@ -10,7 +10,7 @@ import Defaults
 import Luminare
 import SwiftUI
 
-class AdvancedConfigurationModel: ObservableObject {
+final class AdvancedConfigurationModel: ObservableObject {
     @Published private(set) var didImportSuccessfullyAlert = false
     @Published private(set) var didExportSuccessfullyAlert = false
     @Published private(set) var didResetSuccessfullyAlert = false
@@ -155,14 +155,14 @@ struct AdvancedConfigurationView: View {
     }
 
     var body: some View {
-        generalSection()
-        keybindsSection()
-        permissionsSection()
+        generalSection
+        keybindsSection
+        permissionsSection
             .onAppear(perform: model.startTracking)
             .onDisappear(perform: model.stopTracking)
     }
 
-    func generalSection() -> some View {
+    private var generalSection: some View {
         LuminareSection("General") {
             if #available(macOS 15.0, *) {
                 LuminareToggle("Use macOS window manager when available", isOn: $useSystemWindowManagerWhenAvailable)
@@ -210,7 +210,7 @@ struct AdvancedConfigurationView: View {
         }
     }
 
-    func keybindsSection() -> some View {
+    private var keybindsSection: some View {
         LuminareSection("Keybinds") {
             HStack(spacing: 2) {
                 Button(action: model.importPrompt) {
@@ -253,14 +253,14 @@ struct AdvancedConfigurationView: View {
         }
     }
 
-    func permissionsSection() -> some View {
+    private var permissionsSection: some View {
         LuminareSection("Permissions") {
             accessibilityComponent()
         }
         .animation(luminareAnimation, value: model.isAccessibilityAccessGranted)
     }
 
-    func accessibilityComponent() -> some View {
+    private func accessibilityComponent() -> some View {
         LuminareCompose {
             Button {
                 AccessibilityManager.requestAccess()
