+++ Loop/Settings Window/LuminareManager.swift
@@ -10,10 +10,6 @@ import Defaults
 import Luminare
 import SwiftUI
 
-extension String: @retroactive Identifiable {
-    public var id: String { self }
-}
-
 enum Tab: LuminareTabItem, CaseIterable {
     var id: String { title }
 
@@ -78,12 +74,12 @@ enum Tab: LuminareTabItem, CaseIterable {
         }
     }
 
-    static let theming: [Tab] = [.icon, .accentColor, .radialMenu, .preview]
-    static let settings: [Tab] = [.behavior, .keybinds]
-    static let loop: [Tab] = [.advanced, .excludedApps, .about]
+    static let themingTabs: [Tab] = [.icon, .accentColor, .radialMenu, .preview]
+    static let settingsTabs: [Tab] = [.behavior, .keybinds]
+    static let loopTabs: [Tab] = [.advanced, .excludedApps, .about]
 }
 
-class LuminareManager: NSWindowController, ObservableObject {
+final class LuminareManager: NSWindowController, ObservableObject {
     static let shared = LuminareManager()
 
     var luminare: LuminareWindow?
@@ -193,72 +189,6 @@ class LuminareManager: NSWindowController, ObservableObject {
     func stopTimer() {
         previewActionTimerTask?.cancel()
         previewActionTimerTask = nil
-    }
-}
-
-struct LuminareContentView: View {
-    @ObservedObject var model: LuminareManager
-    @ObservedObject private var accentColorController: AccentColorController = .shared
-
-    @Environment(\.luminareAnimation) private var animation
-    @Environment(\.luminareTitleBarHeight) private var titleBarHeight
-
-    var body: some View {
-        LuminareDividedStack {
-            LuminareSidebar {
-                LuminareSidebarSection("Theming", selection: $model.currentTab, items: Tab.theming)
-                LuminareSidebarSection("Settings", selection: $model.currentTab, items: Tab.settings)
-                LuminareSidebarSection("\(Bundle.main.appName)", selection: $model.currentTab, items: Tab.loop)
-            }
-            .frame(width: 260)
-            .padding(.top, titleBarHeight)
-            .luminareBackground()
-
-            LuminarePane {
-                model.currentTab.view()
-            } header: {
-                HStack {
-                    model.currentTab.decoratedImageView
-
-                    Text(model.currentTab.title)
-                        .font(.title2)
-
-                    Spacer()
-
-                    Button {
-                        model.showInspector.toggle()
-                    } label: {
-                        Image(model.showInspector ? .sidebarLeftHide : .sidebarLeft3)
-                    }
-                }
-            }
-            .frame(width: 390)
-
-            if model.showInspector {
-                ZStack {
-                    if model.showPreview {
-                        LuminarePreviewView()
-                    }
-
-                    if model.showRadialMenu {
-                        VStack {
-                            RadialMenuView(viewModel: model.radialMenuViewModel)
-                        }
-                        .frame(maxHeight: .infinity, alignment: .center)
-                    }
-                }
-                .animation(animation, value: [model.showRadialMenu, model.showPreview])
-                .frame(width: 520)
-            }
-        }
-        .onAppear {
-            DispatchQueue.main.async {
-                model.showPreview = true
-                model.showRadialMenu = true
-            }
-        }
-        .luminareTint(overridingWith: accentColorController.color1)
-        .ignoresSafeArea()
     }
 }
