+++ Loop/Settings Window/Settings/Keybindings/Item View/CustomTextField.swift
