+++ Loop/Settings Window/Settings/Keybindings/KeybindsConfigurationView.swift
@@ -19,14 +19,14 @@ struct KeybindsConfigurationView: View {
 
     @StateObject private var model = KeybindsConfigurationModel()
 
-    @Default(.triggerKey) var triggerKey
-    @Default(.triggerDelay) var triggerDelay
-    @Default(.cycleModeRestartEnabled) var cycleModeRestartEnabled
-    @Default(.cycleBackwardsOnShiftPressed) var cycleBackwardsOnShiftPressed
-    @Default(.doubleClickToTrigger) var doubleClickToTrigger
-    @Default(.middleClickTriggersLoop) var middleClickTriggersLoop
-    @Default(.enableTriggerDelayOnMiddleClick) var enableTriggerDelayOnMiddleClick
-    @Default(.keybinds) var keybinds
+    @Default(.triggerKey) private var triggerKey
+    @Default(.triggerDelay) private var triggerDelay
+    @Default(.cycleModeRestartEnabled) private var cycleModeRestartEnabled
+    @Default(.cycleBackwardsOnShiftPressed) private var cycleBackwardsOnShiftPressed
+    @Default(.doubleClickToTrigger) private var doubleClickToTrigger
+    @Default(.middleClickTriggersLoop) private var middleClickTriggersLoop
+    @Default(.enableTriggerDelayOnMiddleClick) private var enableTriggerDelayOnMiddleClick
+    @Default(.keybinds) private var keybinds
 
     /// If the user has "enabled" the trigger delay.
     private var useTriggerDelay: Bool {
@@ -57,84 +57,9 @@ struct KeybindsConfigurationView: View {
 
     var body: some View {
         Group {
-            LuminareSection("Trigger Key") {
-                TriggerKeycorder($triggerKey)
-                    .environmentObject(model)
-                    .luminareBordered(true)
-            }
-            .luminareBordered(false)
-
-            LuminareSection("Settings") {
-                LuminareSlider(
-                    "Trigger delay",
-                    value: $triggerDelay,
-                    in: 0...1,
-                    step: 0.1,
-                    format: .number.precision(.fractionLength(1...1)),
-                    clampsUpper: false,
-                    suffix: .init(.init(localized: "Measurement unit: seconds", defaultValue: "s"))
-                )
-
-                LuminareToggle("Double-click to trigger", isOn: $doubleClickToTrigger)
-                LuminareToggle("Middle-click to trigger", isOn: $middleClickTriggersLoop)
-
-                if showMiddleClickTriggerDelayOption {
-                    LuminareToggle("Apply trigger delay on middle-click", isOn: $enableTriggerDelayOnMiddleClick)
-                }
-
-                if showCycleRestartOption {
-                    LuminareToggle(isOn: $cycleModeRestartEnabled) {
-                        Text("Always start cycles from first item")
-                            .padding(.trailing, 4)
-                            .luminarePopover(attachedTo: .topTrailing) {
-                                Text("By default, Loop resumes cycles from where you last left off in each window.")
-                                    .padding(6)
-                            }
-                    }
-                }
-
-                if showCycleBackwardsOption {
-                    LuminareToggle("Cycle backward with Shift", isOn: $cycleBackwardsOnShiftPressed)
-                }
-            }
-
-            LuminareSection("Keybinds") {
-                HStack(spacing: 2) {
-                    Button("Add") {
-                        keybinds.insert(.init(.noAction), at: 0)
-                    }
-
-                    Button("Remove", role: .destructive) {
-                        keybinds.removeAll(where: model.selectedKeybinds.contains)
-                    }
-                    .disabled(model.selectedKeybinds.isEmpty)
-                    .buttonStyle(.luminareProminent)
-                    .keyboardShortcut(.delete)
-                }
-
-                LuminareList(
-                    items: $keybinds,
-                    selection: $model.selectedKeybinds,
-                    id: \.id
-                ) { keybind in
-                    KeybindItemView(keybind)
-                        .environmentObject(model)
-                } emptyView: {
-                    HStack {
-                        Spacer()
-                        VStack {
-                            Text("No keybinds")
-                                .font(.title3)
-                            Text("Press \"Add\" to add a keybind")
-                                .font(.caption)
-                        }
-                        Spacer()
-                    }
-                    .foregroundStyle(.secondary)
-                    .padding()
-                }
-                .luminareListRoundedCorner(bottom: .always)
-            }
+            triggerKeySection
+            settingsSection
+            keybindsSection
         }
         .animation(
             luminareAnimation,
@@ -144,6 +69,91 @@ struct KeybindsConfigurationView: View {
                 showCycleBackwardsOption
             ]
         )
+    }
+
+    private var triggerKeySection: some View {
+        LuminareSection("Trigger Key") {
+            TriggerKeycorder($triggerKey)
+                .environmentObject(model)
+                .luminareBordered(true)
+        }
+        .luminareBordered(false)
+    }
+
+    private var settingsSection: some View {
+        LuminareSection("Settings") {
+            LuminareSlider(
+                "Trigger delay",
+                value: $triggerDelay,
+                in: 0...1,
+                step: 0.1,
+                format: .number.precision(.fractionLength(1...1)),
+                clampsUpper: false,
+                suffix: .init(.init(localized: "Measurement unit: seconds", defaultValue: "s"))
+            )
+
+            LuminareToggle("Double-click to trigger", isOn: $doubleClickToTrigger)
+            LuminareToggle("Middle-click to trigger", isOn: $middleClickTriggersLoop)
+
+            if showMiddleClickTriggerDelayOption {
+                LuminareToggle("Apply trigger delay on middle-click", isOn: $enableTriggerDelayOnMiddleClick)
+            }
+
+            if showCycleRestartOption {
+                LuminareToggle(isOn: $cycleModeRestartEnabled) {
+                    Text("Always start cycles from first item")
+                        .padding(.trailing, 4)
+                        .luminarePopover(attachedTo: .topTrailing) {
+                            Text("By default, Loop resumes cycles from where you last left off in each window.")
+                                .padding(6)
+                        }
+                }
+            }
+
+            if showCycleBackwardsOption {
+                LuminareToggle("Cycle backward with Shift", isOn: $cycleBackwardsOnShiftPressed)
+            }
+        }
+    }
+
+    private var keybindsSection: some View {
+        LuminareSection("Keybinds") {
+            HStack(spacing: 2) {
+                Button("Add") {
+                    keybinds.insert(.init(.noAction), at: 0)
+                }
+
+                Button("Remove", role: .destructive) {
+                    keybinds.removeAll(where: model.selectedKeybinds.contains)
+                }
+                .disabled(model.selectedKeybinds.isEmpty)
+                .buttonStyle(.luminareProminent)
+                .keyboardShortcut(.delete)
+            }
+
+            LuminareList(
+                items: $keybinds,
+                selection: $model.selectedKeybinds,
+                id: \.id
+            ) { keybind in
+                KeybindItemView(keybind)
+                    .environmentObject(model)
+            } emptyView: {
+                HStack {
+                    Spacer()
+                    VStack {
+                        Text("No keybinds")
+                            .font(.title3)
+                        Text("Press \"Add\" to add a keybind")
+                            .font(.caption)
+                    }
+                    Spacer()
+                }
+                .foregroundStyle(.secondary)
+                .padding()
+            }
+            .luminareListRoundedCorner(bottom: .always)
+        }
     }
 }
